Move tc/core/libraries.h to tc/core/cuda/cuda_libraries.h
This was left out of the commit in prehistory that moved CUDA-specific
files to cuda/ subdirectories.
